Fix FSTDEP006 by exposing the Authenticator object using a getter decorator
We're decorating the request with the Authenticator object responsible for managing passport authentication mostly so that request handlers installed by the Authenticator can ensure they're operating in an encapsulation context where the rest of the fastify-passport plugin was initialized correctly. Fastify has deprecated reference decorators to avoid accidental data-sharing between requests or memory leaks. In this case, I think we actually *want* to share this piece of data (the Authenticator instance) between requests, so we use the escape hatch to avoid the deprecation warning by using a getter decorator to return it.

An alternative might be to set a scalar sigil of some sort on the request, but I feel like that's less clear, or to use an onRequest hook, which is slower just to avoid the deprecation warning, which I think is silly. So, a getter it is!