Implement and document handled deserialization failures
I missed this nuance during the original porting -- returning false or null from a user deserializer means something special. It means the user not being found is not an error condition but just means the session has out-of-date data, and should be logged out. See https://github.com/jaredhanson/passport/issues/6 for the original context on how this is a thing in passport. This recreates that functionality and tests it!