Add support for authenticating with already-instantiated strategies.
This allows for running authentication with ephemeral strategies you don't want to register into the global namespace, or for simultaneous asynchronous authentications using the same strategy with different configurations. My use case specifically is that I want some users of my application to bring their own credentials -- instead of using one fixed oauth access token for a given integration, users should be able to supply their own, and we'll use it at runtime instead of the fixed one.

This means that for a given request we need to be able to use either the fixed strategy registered with the authenticator at server boot, or an ephemeral one created using the request context for just that request. It's important that the strategy not be registered into the global list of strategies so it doesn't stomp on the strategy with the same name, or leak memory if given a different name.

Most of this change is just TypeScript types and tests -- only one small implementation change to only sometimes instantiate the sent in strategies by name.