Ensure current behaviour of late-binding strategies by name is preserved
Before the changes in #222, strategies were looked up in the global list by name at route invocation time, which means you can create authentication handlers that refer to strategies by name before actually registering those strategies. This is kinda slow, but more flexible, and is already out in the wild in passport and fastify-passport, so I think we should preserve this behaviour.