Catch unexpected errors during strategy execution and dispatch through the fastify error stack
Before this change, strategies that throw errors (or return promises which throw errors) wouldn't trigger the fastify error handling. Strategies which returned promises that then rejected would trigger uncaughPromiseRejection events as well.

This issue stems from the strange API that passport requires for strategies born in the callback era, where strategies get told to run and are expected to call back into the fastify-passport handling by calling `this.error` or `this.success` or what have you. If the strategy throws unexpectedly though, we didn't catch that and now we do.